test: Корректные и некорректные тестовые кейсы для visible_string feat: Геттер значения для базового класса строковых типов test:Тесты visible_string test:Тесты bmp_string feat: Определен шаблон для bmp_string

+++ include/asncpp/visible_string.h
@@ -7,6 +7,6 @@
 #include "base/string_base.h"
 #include "base/asn_base.h"
 
-using visible_string_t [[deprecated]] = string_base_t<char, asn1_tag::VISIBLE_STRING, [](char c) { return std::isprint(c); }>;
+using visible_string_t [[deprecated]] = string_base_t<char, asn1_tag::VISIBLE_STRING, [](char c) { return std::isprint(static_cast<uint8_t>(c)); }>;
 
 #endif //VISIBLE_STRING_H
